+++ app/src/main/java/eu/wedgess/piholecontrol/data/model/responses/PiHoleOverTimeData.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PiHoleOverTimeData(
     @SerialName("domains_over_time")
-    val domainsOverTime: Map<Float, Float> = mapOf(),
+    val domainsOverTime: Map<Long, Long> = mapOf(),
     @SerialName("ads_over_time")
-    val adsOverTime: Map<Float, Float> = mapOf(),
+    val adsOverTime: Map<Long, Long> = mapOf(),
 )
